rename RAC Accordion file to Disclosure
